+++ examples/stats/plot_oneWay_Manova.py
@@ -5,7 +5,10 @@ One Way manova
 
 One way manova to compare Left vs Right.
 """
-import matplotlib.pyplot as plt
+import seaborn as sns
+
+from time import time
+from pylab import plt
 
 from mne import Epochs, pick_types
 from mne.io import concatenate_raws
@@ -13,9 +16,14 @@ from mne.io.edf import read_raw_edf
 from mne.datasets import eegbci
 from mne.event import find_events
 
-from pyriemann.stats import PermutationTest
+from pyriemann.stats import PermutationDistance, PermutationModel
 from pyriemann.estimation import Covariances
+from pyriemann.spatialfilters import CSP
 
+from sklearn.pipeline import make_pipeline
+from sklearn.linear_model import LogisticRegression
+
+sns.set_style('whitegrid')
 ###############################################################################
 # Set parameters and read data
 
@@ -36,6 +44,7 @@ raw.filter(7., 35., method='iir')
 events = find_events(raw, shortest_event=0, stim_channel='STI 014')
 picks = pick_types(raw.info, meg=False, eeg=True, stim=False, eog=False,
                    exclude='bads')
+picks = picks[::4]
 
 # Read epochs (train will be done only between 1 and 2s)
 # Testing will be done with a running classifier
@@ -48,10 +57,63 @@ epochs_data = epochs.get_data()
 
 # compute covariance matrices
 covmats = Covariances().fit_transform(epochs_data)
+fig, axes = plt.subplots(2, 2, figsize=[12, 6], sharey=True)
 
-p_test = PermutationTest(5000)
+###############################################################################
+# Pairwise distance based permutation test
+###############################################################################
+
+t_init = time()
+p_test = PermutationDistance(500, metric='riemann', mode='pairwise')
 p, F = p_test.test(covmats, labels)
-p_test.plot()
-print(p_test.summary())
+duration = time() - t_init
 
+p_test.plot(axes=axes[0, 0])
+axes[0, 0].set_title('Pairwise distance - %.2f sec.' % duration)
+print('p-value: %.3f' % p)
+
+###############################################################################
+# t-test distance based permutation test
+###############################################################################
+
+t_init = time()
+p_test = PermutationDistance(500, metric='riemann', mode='ttest')
+p, F = p_test.test(covmats, labels)
+duration = time() - t_init
+
+p_test.plot(axes=axes[0, 1])
+axes[0, 1].set_title('t-test distance - %.2f sec.' % duration)
+print('p-value: %.3f' % p)
+
+###############################################################################
+# F-test distance based permutation test
+###############################################################################
+
+t_init = time()
+p_test = PermutationDistance(500, metric='riemann', mode='ftest')
+p, F = p_test.test(covmats, labels)
+duration = time() - t_init
+
+p_test.plot(axes=axes[1, 0])
+axes[1, 0].set_title('F-test distance - %.2f sec.' % duration)
+print('p-value: %.3f' % p)
+
+
+###############################################################################
+# Classification based permutation test
+###############################################################################
+
+clf = make_pipeline(CSP(4), LogisticRegression())
+
+t_init = time()
+p_test = PermutationModel(500, model=clf, cv=3)
+p, F = p_test.test(covmats, labels)
+duration = time() - t_init
+
+p_test.plot(axes=axes[1, 1])
+axes[1, 1].set_title('Classification - %.2f sec.' % duration)
+print('p-value: %.3f' % p)
+
+sns.despine()
+plt.tight_layout()
 plt.show()