+++ pyriemann/utils/distance.py
@@ -134,6 +134,25 @@ def distance(A, B, metric='riemann'):
     return d
 
 
+def pairwise_distance(X, Y=None, metric='riemann'):
+    """Pairwise distance matrix"""
+    Ntx, _, _ = X.shape
+
+    if Y is None:
+        dist = numpy.zeros((Ntx, Ntx))
+        for i in range(Ntx):
+            for j in range(i + 1, Ntx):
+                dist[i, j] = distance(X[i], X[j], metric)
+        dist += dist.T
+    else:
+        Nty, _, _ = Y.shape
+        dist = numpy.empty((Ntx, Nty))
+        for i in range(Ntx):
+            for j in range(Nty):
+                dist[i, j] = distance(X[i], Y[j], metric)
+    return dist
+
+
 distance_methods = {'riemann': distance_riemann,
                     'logeuclid': distance_logeuclid,
                     'euclid': distance_euclid,